requirements.txt
pyTelegramBotAPI
torch
torchvision
pillow
git+https://github.com/openai/CLIP.git
